+++ awb_amount_based_approvals/__manifest__.py
@@ -21,12 +21,13 @@
 
     'category': 'Approvals',
 
-    'version': '13.0.1.1.0',
+    'version': '13.0.1.1.1',
 
-    'depends': ['approvals'],
+    'depends': ['approvals', 'requisition_approval'],
 
     'data': [
         'security/ir.model.access.csv',
+        'security/ir.rule.csv',
         'views/approval_rules_view.xml',
         'views/approval_types_view.xml',
         'views/approval_request_view.xml',