+++ awb_payment_allocation/models/__init__.py
@@ -7,3 +7,4 @@
 
 # from . import account
 from . import account_payment
+from . import payment_allocation