+++ awb_product_segmentation/CHANGELOG.md
@@ -1,2 +1,5 @@
+## [13.0.1.0.1] - 2021-08-24
+- [ADD] Logo and Fix Manifest Summary
+
 ## [13.0.1.0.0] - 2021-07-29
 - [ADD] Product Segmentation