+++ bill-payment/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.3.1] - 2021-08-24
+
+- [FIX] Recompute Statement Line Date Cut-offs 
+
+## [13.0.1.3.0] - 2021-08-23
+
+- [IMP] Recompute Statement Line
+
 ## [13.0.1.2.0] - 2021-04-22
 
 - [IMP] Total of Multiple Rebates